Fixed setEntryDelay and setExitDelay
I've updated the setEntryDelay and setExitDelay function so that it works properly with new Lock Manager. I noticed that when the function input was defined as "delay = 30" it would not properly take the delay sent from a SmartApp. I've made these changes in my copy of the DTH and it works as expected.